fininfocom-employee/requirements.txt
asgiref==3.7.2
Django==3.2.23
django-filter==23.4
djangorestframework==3.14.0
importlib-metadata==6.7.0
Markdown==3.4.4
Pillow==9.5.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.7.1
zipp==3.15.0
